feat(policies): improve customability of policy context
