+++ packages/policies/tests/setup/policies/index.ts
@@ -1,2 +1,1 @@
-export * from './posts/edit-post.policy'
-export * from './posts/view-post.policy'
+export * from './posts.policy'
